pyproject.toml
[project]
name = "expo-wait-times-visualization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash>=3.2.0",
    "dash-bootstrap-components>=2.0.3",
    "dash-leaflet>=0.1.28",
    "flask-caching>=2.3.0",
    "plotly>=5.22",
    "pandas>=2.3.1",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "dash-extensions>=1.0.14",
    "dash-leaflet>=0.1.28",
]